requirements.txt
python-dotenv==1.0.1
click==8.1.8
requests==2.32.3
polars==1.5.0
numpy==1.24.4
fastsparsegams==0.1.2